fix: align container sarif output
