+++ .github/CODEOWNERS
@@ -21,19 +21,21 @@ src/cli/commands/describe.ts @snyk/cloud-context
 src/cli/commands/update-exclude-policy.ts @snyk/cloud-dev-ex
 src/cli/commands/apps @snyk/moose
 src/lib/apps @snyk/moose
-src/lib/container @snyk/mycelium
+src/lib/container @snyk/lumos
 src/lib/plugins @snyk/snyk-open-source
 test/fixtures/sast/ @snyk/zenith
 src/lib/plugins/sast/ @snyk/zenith
 test/jest/unit/snyk-code/ @snyk/zenith
 src/lib/formatters/iac-output/ @snyk/cloud-dev-ex
+src/lib/formatters/sarif-output.ts @snyk/container
+src/lib/formatters/get-sarif-result.ts @snyk/container @snyk/hammerhead @snyk/snyk-open-source
 src/lib/iac/ @snyk/group-infrastructure-as-code
 src/lib/iac/test/ @snyk/cloud-dev-ex
 src/lib/snyk-test/iac-test-result.ts @snyk/cloud-dev-ex
-test/fixtures/basic-apk/ @snyk/mycelium
-test/fixtures/container-app-vulns/ @snyk/mycelium
-test/fixtures/container-projects/ @snyk/mycelium @snyk/potion
-test/fixtures/docker/ @snyk/mycelium @snyk/potion
+test/fixtures/basic-apk/ @snyk/lumos
+test/fixtures/container-app-vulns/ @snyk/lumos
+test/fixtures/container-projects/ @snyk/lumos
+test/fixtures/docker/ @snyk/lumos
 test/fixtures/iac/ @snyk/cloud-dev-ex
 test/fixtures/iac/drift @snyk/cloud-context
 test/fixtures/iac/capture @snyk/cloud-context
@@ -43,7 +45,7 @@ test/smoke/spec/snyk_basic_spec.sh @snyk/hammerhead
 test/smoke/.iac-data/ @snyk/cloud-dev-ex
 test/jest/unit/lib/endpoint-config-test.spec.ts @snyk/nebula
 test/jest/unit/lib/formatters/iac-output/ @snyk/cloud-dev-ex
-test/jest/unit/lib/formatters/test/format-test-results.spec.ts @snyk/hammerhead @snyk/snyk-open-source @snyk/mycelium
+test/jest/unit/lib/formatters/test/format-test-results.spec.ts @snyk/hammerhead @snyk/snyk-open-source @snyk/lumos
 test/jest/unit/iac/ @snyk/cloud-dev-ex
 test/jest/unit/cli/commands/test/iac @snyk/cloud-dev-ex
 test/jest/unit/lib/iac/drift/ @snyk/cloud-context
@@ -72,7 +74,7 @@ src/cli/commands/log4shell-hashes.ts @snyk/tundra
 src/cli/commands/log4shell.ts @snyk/tundra
 test/fixtures/unmanaged-log4j-fixture @snyk/tundra
 test/jest/acceptance/snyk-log4shell/log4shell-detection.spec.ts @snyk/tundra
-test/jest/acceptance/snyk-test/app-vuln-container-project.spec.ts @snyk/mycelium
+test/jest/acceptance/snyk-test/app-vuln-container-project.spec.ts @snyk/lumos
 /.github @snyk/hammerhead
 /.github/workflows/iac-smoke-tests.yml @snyk/cloud-dev-ex
 /.github/workflows/iac-smoke-tests-pulls.yml @snyk/cloud-dev-ex
