+++ src/lib/formatters/open-source-sarif-output.ts
@@ -1,9 +1,10 @@
 import * as sarif from 'sarif';
-const upperFirst = require('lodash.upperfirst');
-const groupBy = require('lodash.groupby');
-const map = require('lodash.map');
+import * as upperFirst from 'lodash.upperfirst';
+import * as groupBy from 'lodash.groupby';
+import * as map from 'lodash.map';
 
-import { TestResult, SEVERITY, AnnotatedIssue } from '../snyk-test/legacy';
+import { TestResult, AnnotatedIssue } from '../snyk-test/legacy';
+import { getResults } from './get-sarif-result';
 
 const LOCK_FILES_TO_MANIFEST_MAP = {
   'Gemfile.lock': 'Gemfile',
@@ -87,45 +88,6 @@ ${vuln.description}`.replace(/##\s/g, '# '),
       };
     },
   );
-}
-
-export function getResults(testResult): sarif.Result[] {
-  const groupedVulnerabilities = groupBy(testResult.vulnerabilities, 'id');
-  return map(
-    groupedVulnerabilities,
-    ([vuln]): sarif.Result => ({
-      ruleId: vuln.id,
-      level: getLevel(vuln),
-      message: {
-        text: `This file introduces a vulnerable ${vuln.packageName} package with a ${vuln.severity} severity vulnerability.`,
-      },
-      locations: [
-        {
-          physicalLocation: {
-            artifactLocation: {
-              uri: testResult.displayTargetFile,
-            },
-            region: {
-              startLine: vuln.lineNumber || 1,
-            },
-          },
-        },
-      ],
-    }),
-  );
-}
-
-export function getLevel(vuln: AnnotatedIssue) {
-  switch (vuln.severity) {
-    case SEVERITY.CRITICAL:
-    case SEVERITY.HIGH:
-      return 'error';
-    case SEVERITY.MEDIUM:
-      return 'warning';
-    case SEVERITY.LOW:
-    default:
-      return 'note';
-  }
 }
 
 function getIntroducedThrough(vuln: AnnotatedIssue) {
