+++ src/lib/formatters/sarif-output.ts
@@ -1,12 +1,15 @@
 import * as sarif from 'sarif';
+import * as upperFirst from 'lodash.upperfirst';
 import { TestResult } from '../snyk-test/legacy';
 import { SEVERITY } from '../snyk-test/legacy';
-const upperFirst = require('lodash.upperfirst');
+import { getResults } from './get-sarif-result';
 
 export function createSarifOutputForContainers(
   testResults: TestResult[],
 ): sarif.Log {
   const sarifRes: sarif.Log = {
+    $schema:
+      'https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json',
     version: '2.1.0',
     runs: [],
   };
@@ -48,7 +51,7 @@ export function getTool(testResult): sarif.Tool {
         return;
       }
       const level = getIssueLevel(vuln.severity);
-      const cve = vuln['identifiers']['CVE'][0];
+      const cve = vuln.identifiers?.CVE?.join();
       pushedIds[vuln.id] = true;
       return {
         id: vuln.id,
@@ -70,39 +73,15 @@ export function getTool(testResult): sarif.Tool {
           level: level,
         },
         properties: {
-          tags: ['security', ...vuln.identifiers.CWE],
+          tags: [
+            'security',
+            ...(vuln.identifiers?.CWE || []),
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            testResult.packageManager!,
+          ],
         },
       };
     })
     .filter(Boolean);
   return tool;
-}
-
-export function getResults(testResult): sarif.Result[] {
-  const results: sarif.Result[] = [];
-
-  if (!testResult.vulnerabilities) {
-    return results;
-  }
-  testResult.vulnerabilities.forEach((vuln) => {
-    results.push({
-      ruleId: vuln.id,
-      message: {
-        text: `This file introduces a vulnerable ${vuln.packageName} package with a ${vuln.severity} severity vulnerability.`,
-      },
-      locations: [
-        {
-          physicalLocation: {
-            artifactLocation: {
-              uri: testResult.displayTargetFile,
-            },
-            region: {
-              startLine: vuln.lineNumber || 1,
-            },
-          },
-        },
-      ],
-    });
-  });
-  return results;
 }
