+++ test/fixtures/docker/sarif-container-result.json
@@ -1,4 +1,5 @@
 {
+  "$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
   "version": "2.1.0",
   "runs": [
     {
@@ -18,8 +19,15 @@
                 "text": "",
                 "markdown": "## Overview\nUse-after-free vulnerability in bzip2recover in bzip2 1.0.6 allows remote attackers to cause a denial of service (crash) via a crafted bzip2 file, related to block ends set to before the start of the block.\n\n## References\n- [GENTOO](https://security.gentoo.org/glsa/201708-08)\n- [CONFIRM](https://bugzilla.redhat.com/show_bug.cgi?id=1319648)\n- [SECTRACK](http://www.securitytracker.com/id/1036132)\n- [BID](http://www.securityfocus.com/bid/91297)\n- [CONFIRM](http://www.oracle.com/technetwork/topics/security/bulletinjul2016-3090568.html)\n- [MLIST](http://www.openwall.com/lists/oss-security/2016/06/20/1)\n"
               },
-              "defaultConfiguration": { "level": "warning" },
-              "properties": { "tags": ["security"] }
+              "defaultConfiguration": {
+                "level": "warning"
+              },
+              "properties": {
+                "tags": [
+                  "security",
+                  "deb"
+                ]
+              }
             }
           ]
         }
@@ -27,6 +35,7 @@
       "results": [
         {
           "ruleId": "SNYK-LINUX-BZIP2-106947",
+          "level": "note",
           "message": {
             "text": "This file introduces a vulnerable bzip2 package with a low severity vulnerability."
           },
@@ -34,21 +43,9 @@
             {
               "physicalLocation": {
                 "artifactLocation": {},
-                "region": { "startLine": 1 }
-              }
-            }
-          ]
-        },
-        {
-          "ruleId": "SNYK-LINUX-BZIP2-106947",
-          "message": {
-            "text": "This file introduces a vulnerable bzip2 package with a low severity vulnerability."
-          },
-          "locations": [
-            {
-              "physicalLocation": {
-                "artifactLocation": {},
-                "region": { "startLine": 1 }
+                "region": {
+                  "startLine": 1
+                }
               }
             }
           ]
